requirements.txt
torch
scikit-image
numpy
matplotlib
argparse
ipython
pillow
opencv-python
tkvideo